% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_Test.R
\name{pre.test}
\alias{pre.test}
\title{pre.test() — Automatic statistical test suggestion}
\usage{
pre.test(..., alpha = 0.05, help = FALSE, verbose = TRUE)
}
\arguments{
\item{...}{Two or more vectors (numeric or categorical), or a data.frame with >= 2 columns}

\item{alpha}{Significance level. Default = 0.05}

\item{help}{Logical. If TRUE, shows detailed help}

\item{verbose}{Logical. If TRUE, prints informative messages}
}
\value{
Invisible list with normality results, homogeneity or contingency table,
and test recommendation
}
\description{
Automatically identifies whether the input data are numeric or categorical
and suggests the most appropriate statistical test.
}
