% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA_test.R
\name{test.anova}
\alias{test.anova}
\title{ANOVA test with automated assumption checking}
\usage{
test.anova(
  ...,
  title = "ANOVA/Tukey HSD",
  x = "X axis",
  y = "Y axis",
  style = 1,
  help = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{...}{Vectors or a data.frame with >= 2 columns.}

\item{title}{Plot title.}

\item{x}{X-axis label.}

\item{y}{Y-axis label.}

\item{style}{Aesthetic style of the generated plot.}

\item{help}{If TRUE, shows help.}

\item{verbose}{If TRUE, shows detailed messages.}
}
\value{
An \code{aov} object or a recommendation message.
}
\description{
Performs ANOVA (and Tukey HSD) if data meet normality and homogeneity assumptions.
Otherwise, automatically recommends Kruskal-Wallis/Dunn.
}
\examples{
df <- data.frame(
  control = rnorm(30, 10),
  treatment = rnorm(30, 12),
  test = rnorm(30, 11)
)
test.anova(df)
}
