% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlation_test.R
\name{test.correlation}
\alias{test.correlation}
\title{Correlation test (Pearson, Spearman or Kendall)}
\usage{
test.correlation(
  x,
  y = NULL,
  method = "auto",
  style = 1,
  help = FALSE,
  verbose = TRUE,
  plot_normality = FALSE
)
}
\arguments{
\item{x}{Numeric vector or data frame with two numeric columns}

\item{y}{Numeric vector (optional if x is a data frame)}

\item{method}{Correlation method: "auto" (default), "pearson", "spearman" or "kendall"}

\item{style}{Plot aesthetic style.}

\item{help}{If TRUE, displays a detailed explanation of the function}

\item{verbose}{If TRUE, prints messages about the chosen method and normality tests}

\item{plot_normality}{If TRUE, generates QQ-plots to assess data normality}
}
\value{
Invisible object of class htest with the correlation test result
}
\description{
Performs correlation testing between two numeric variables, automatically choosing
between Pearson, Spearman or Kendall based on normality, ties and outliers.
Optionally displays diagnostic plots and a correlation plot with trend line.
}
\examples{
# Pearson
set.seed(123)
x <- rnorm(50, sd = 0.1)
y <- x + rnorm(50, sd = 0.1)
test.correlation(x, y)

# Spearman
set.seed(123)
x <- runif(300)
y <- log(x + 0.1) + rnorm(300, sd = 0.5)
test.correlation(x, y)

# Kendall
set.seed(123)
x <- runif(1000, 1, 100)
y <- sin(x) * 30 + rnorm(1000, 0, 10)
x[sample(1:500, 50)] <- 50
y[sample(1:500, 50)] <- 0
x_out <- runif(100, 10, 20)
y_out <- runif(100, 80, 120)
x <- c(x, x_out)
y <- c(y, y_out)
test.correlation(x, y)

}
