% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kruskal_test.R
\name{test.kruskal}
\alias{test.kruskal}
\title{Kruskal-Wallis Test with Dunn Post-hoc}
\usage{
test.kruskal(
  ...,
  title = "Kruskal-Wallis + Dunn",
  x_label = "Group",
  y_label = "Value",
  style = 1,
  help = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{...}{Numeric vectors or a data frame with two or more columns (each representing a group)}

\item{title}{Plot title (default = "Kruskal-Wallis + Dunn")}

\item{x_label}{X-axis label (default = "Group")}

\item{y_label}{Y-axis label (default = "Value")}

\item{style}{Plot style.}

\item{help}{Logical. If TRUE, shows help message}

\item{verbose}{Logical. If TRUE, prints detailed messages (default = TRUE)}
}
\value{
Kruskal-Wallis test object (invisible)
}
\description{
Performs the Kruskal-Wallis test for comparing multiple independent groups,
followed by Dunn's post-hoc test with Bonferroni correction.
}
